backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.1.1
httpx==0.27.2
python-multipart==0.0.12
orjson==3.10.7
